Move the sync code out of the frontend/flight module into it's own module
