+++ src/lib.rs
@@ -12,6 +12,7 @@ pub mod nodes;
 pub mod object_store;
 pub mod provider;
 pub mod repository;
+pub mod sync;
 pub mod system_tables;
 pub mod utils;
 pub mod version;
